requirements.txt
Django==1.11.6
django-allauth==0.34.0
django-bootstrap3==9.1.0
django-ckeditor==5.4.0
django-crispy-forms==1.7.0
django-extensions==1.9.1
django-js-asset==0.1.1
django-markdown-deux==1.0.5
django-pagedown==1.0.4
djangorestframework==3.7.1
markdown2==2.3.5
oauthlib==1.0.3
python3-openid==3.1.0
requests==2.9.1
requests-oauthlib==0.8.0
urllib3==1.13.1
pillow==5.0.0
django-rest-auth==0.9.3
